+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for NeoFS Node
 ### Changed
 - IR calls `ObjectService.SearchV2` to select SG objects now (#3144)
 - IR contract approval no longer supports neofsid accounts (#3256)
+- IR no longer processes bind/unbind neofsid requests (#3256)
 
 ### Removed